Cleaned up count down mechanism and fixed reset()
Now, any change to the barrier updates the time to stability, and that is displayed on the EzRMS canvas.

After Ez stability is reached, a new countdown starts for 10*tau/dt time steps. Averaging Ez^2 takes place during this time.

Fast forward now speeds to the end of whichever countdown is active.